Bug 1674777 part 5 - Change GetObjectAs* APIs to use size_t for outparam instead of uint32_t. r=sfink,sg
Note that in the IndexedDB code we pass a Span to Key::EncodeAsString where we check
for large lengths.

Differential Revision: https://phabricator.services.mozilla.com/D103679

+++ js/public/ArrayBuffer.h
@@ -197,7 +197,7 @@ extern JS_PUBLIC_API JSObject* UnwrapArrayBuffer(JSObject* obj);
  * |*data|.
  */
 extern JS_PUBLIC_API JSObject* GetObjectAsArrayBuffer(JSObject* obj,
-                                                      uint32_t* length,
+                                                      size_t* length,
                                                       uint8_t** data);
 
 /**
